% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearn.R
\name{yearn}
\alias{yearn}
\title{Do yearn on a vector}
\usage{
yearn(package, maxdist = 0)
}
\arguments{
\item{package}{A vector of packages (could be length 1)}

\item{maxdist}{The maximum distance that counts as a match}
}
\value{
A list of return strings (NULL if installed); by default invisible, wrap in parens if you want to see
}
\description{
Do yearn on a vector
}
