% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearn.R
\name{yearn.one}
\alias{yearn.one}
\title{Do yearn on a single package}
\usage{
yearn.one(pkg)
}
\arguments{
\item{pkg}{The string with the package name}
}
\value{
NULL if success, a string describing the problem ohterwise
}
\description{
Do yearn on a single package
}
