% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearn.R
\name{yearn}
\alias{yearn}
\title{Do yearn on a vector}
\usage{
yearn(package, success.return = FALSE)
}
\arguments{
\item{package}{A vector of packages (could be length 1)}

\item{success.return}{If FALSE, only return info on failing packages}
}
\value{
A list of return strings (NULL if installed)
}
\description{
Do yearn on a vector
}
