% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearn.R
\name{yearn}
\alias{yearn}
\title{Do yearn on a single package}
\usage{
yearn(pkg, maxdist = 2, username.pref = c("cran", "ropensci", "rstudio",
  "tidyverse", "hadley", "yihui", "RcppCore", "eddelbuettel", "ropenscilabs",
  "hrbrmstr", "thej022214", "bomeara"))
}
\arguments{
\item{pkg}{The package name (bare text, not in quotes)}

\item{maxdist}{The maximum distance that counts as a match}

\item{username.pref}{In case of matches, user names in order of preference.}
}
\value{
NULL if success, a string describing the problem otherwise
}
\description{
Do yearn on a single package
}
